+++ packages/redux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `Response` type replaces `AnyResponse`.
+
 ## 1.0.0-alpha.2
 
 * Re-add Curi config reducer and syncing in `syncResponses`. This will allow the user to grab everything from their store instead of having to use `curious` to get the values off of React's `context`.